feat : context 삭제
