+++ src/components/IssueList.tsx
@@ -1,44 +1,27 @@
-import { Fragment } from 'react';
-import { useNavigate } from 'react-router-dom';
 import { styled } from 'styled-components';
 
-import { useIssueContext } from '@/context/useIssueContext';
-import { Advertise } from './Advertise';
-import { IssueInfo } from './IssueInfo';
-
-function isAdCell(index: number) {
-  return index !== 0 && index % 4 === 0;
-}
+import { PER_PAGE } from '@/api/issue';
+import { useFetchIssues } from '@/hook/useFetchIssues';
+import { useInfiniteScroll } from '@/hook/useInfiniteScroll';
+import { Loading } from './common/Loading';
+import { Issues } from './Issues';
 
 export function IssueList() {
-  const issues = useIssueContext();
-  const navigate = useNavigate();
-
-  const handleClickIssue = (number: number) => navigate(`${number}`);
+  const [issues, isLoading, hasNextPage, fetchNextPage] = useFetchIssues(PER_PAGE);
+  const ref = useInfiniteScroll(async (entry, observer) => {
+    observer.unobserve(entry.target);
+    if (!isLoading && hasNextPage) fetchNextPage();
+  });
 
   return (
-    <>
-      {issues.map((issue, index) => (
-        <Fragment key={issue.id}>
-          {isAdCell(index) && <Advertise />}
-          <Box onClick={() => handleClickIssue(issue.number)}>
-            <IssueInfo
-              issueNumber={issue.number}
-              title={issue.title}
-              author={issue.user?.login}
-              created_at={issue.created_at}
-              comments={issue.comments}
-            />
-          </Box>
-        </Fragment>
-      ))}
-    </>
+    <div>
+      <Issues issues={issues} />
+      <Target ref={ref} />
+      {isLoading && <Loading />}
+    </div>
   );
 }
 
-const Box = styled.div`
-  padding: 20px;
-  box-shadow: 0px 0px 5px 0px #bcbcbc;
-  margin-bottom: 20px;
-  cursor: pointer;
+const Target = styled.div`
+  height: 1px;
 `;
