--- src/context/IssueContext.tsx
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-import { IssueData } from '@/api/issue';
-
-export const IssueContext = createContext<IssueData[]>([]);
