+++ src/hook/useFetchIssues.tsx
@@ -5,7 +5,7 @@ import { fetchGithubIssues, IssueData } from '@/api/issue';
 export type InfiniteQuery = [data: IssueData[], isLoading: boolean, hasNextPage: boolean, fetchNextPage: () => void];
 
 export function useFetchIssues(size: number): InfiniteQuery {
-  const [page, setPage] = useState(0);
+  const [page, setPage] = useState(1);
   const [hasNextPage, setHasNextPage] = useState(true);
   const [isLoading, setIsLoading] = useState(true);
   const [data, setData] = useState<IssueData[]>([]);
@@ -13,23 +13,14 @@ export function useFetchIssues(size: number): InfiniteQuery {
   const fetchNextPage = () => setPage((page) => page + 1);
 
   useEffect(() => {
-    fetchGithubIssues().then((res) => {
+    setIsLoading(true);
+
+    fetchGithubIssues(page).then((res) => {
       if (res.length < size) setHasNextPage(false);
-      setData(res);
-      setPage(1);
+
+      setData((prev) => [...prev, ...res]);
       setIsLoading(false);
     });
-  }, [size]);
-
-  useEffect(() => {
-    if (page > 1) {
-      setIsLoading(true);
-      fetchGithubIssues(page).then((res) => {
-        if (res.length < size) setHasNextPage(false);
-        setData((prev) => [...prev, ...res]);
-        setIsLoading(false);
-      });
-    }
   }, [page, size]);
 
   return [data, isLoading, hasNextPage, fetchNextPage];
