+++ src/routes.tsx
@@ -2,8 +2,8 @@ import { Suspense } from 'react';
 import { createBrowserRouter, Navigate } from 'react-router-dom';
 
 import { Loading } from './components/common/Loading';
-import { IssueContent } from './components/IssueContent';
-import { IssueProvider } from './components/IssueProvider';
+import { Issue } from './components/Issue';
+import { IssueList } from './components/IssueList';
 import { ErrorPage } from './pages/ErrorPage';
 import { IssuePage } from './pages/IssuePage';
 
@@ -18,13 +18,13 @@ export const router = createBrowserRouter([
     children: [
       {
         path: '',
-        element: <IssueProvider />,
+        element: <IssueList />,
       },
       {
         path: ':number',
         element: (
           <Suspense fallback={<Loading />}>
-            <IssueContent />
+            <Issue />
           </Suspense>
         ),
       },
